+++ pytorch_lightning/callbacks/progress/progress.py
@@ -11,12 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from typing import Any
+
 from pytorch_lightning.callbacks.progress.tqdm_progress import TQDMProgressBar
 from pytorch_lightning.utilities import rank_zero_deprecation
 
 
 class ProgressBar(TQDMProgressBar):
-    def __init__(self, *args, **kwargs) -> None:
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
         super().__init__(*args, **kwargs)
         rank_zero_deprecation(
             "`ProgressBar` has been deprecated in v1.5 and will be removed in v1.7."