+++ pytorch_lightning/strategies/strategy.py
@@ -381,7 +381,7 @@ class Strategy(ABC):
         return False
 
     @property
-    def lightning_restore_optimizer_and_schedulers(self) -> bool:
+    def lightning_restore_optimizer(self) -> bool:
         """Override to disable Lightning restoring optimizers/schedulers.
 
         This is useful for plugins which manage restoring optimizers/schedulers.